+++ chartlets.py/demo/my_extension/__init__.py
@@ -11,6 +11,7 @@ from .my_panel_5 import panel as my_panel_5
 from .my_panel_6 import panel as my_panel_6
 from .my_panel_7 import panel as my_panel_7
 from .my_panel_8 import panel as my_panel_8
+from .my_panel_9 import panel as my_panel_9
 
 
 ext = Extension(__name__)
@@ -22,3 +23,4 @@ ext.add(my_panel_5)
 ext.add(my_panel_6)
 ext.add(my_panel_7)
 ext.add(my_panel_8)
+ext.add(my_panel_9)