+++ chartlets.js/CHANGES.md
@@ -8,6 +8,8 @@
 
 * Added icon support for `Button`, `IconButton` and `Tabs` components.
  (#124).
+ 
+* Added (MUI) component `Accordion`. (#134)
 
 ## Version 0.1.7 (from 2025/12/03)